Phase class relocated to org.incha.core.telemetry.utils package
