+++ jswingripples/src/main/java/org/incha/core/telemetry/AbstractTelemetryLogger.java
@@ -1,5 +1,7 @@
 package org.incha.core.telemetry;
 
+import org.incha.core.telemetry.utils.Phase;
+
 import java.text.SimpleDateFormat;
 import java.util.Calendar;
