+++ jswingripples/src/main/java/org/incha/core/telemetry/utils/CSVLogSerializer.java
@@ -1,7 +1,5 @@
 package org.incha.core.telemetry.utils;
 
-import org.incha.core.telemetry.Phase;
-
 import java.io.*;
 
 /**
